+++ pmd-core/src/main/java/net/sourceforge/pmd/RuleSet.java
@@ -21,7 +21,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import net.sourceforge.pmd.annotation.InternalApi;
-import net.sourceforge.pmd.cache.ChecksumAware;
+import net.sourceforge.pmd.cache.internal.ChecksumAware;
 import net.sourceforge.pmd.internal.util.PredicateUtil;
 import net.sourceforge.pmd.lang.LanguageVersion;
 import net.sourceforge.pmd.lang.document.FileId;