+++ pmd-core/src/main/java/net/sourceforge/pmd/cache/internal/ChecksumAware.java
@@ -2,9 +2,7 @@
  * BSD-style license; for more info see http://pmd.sourceforge.net/license.html
  */
 
-package net.sourceforge.pmd.cache;
-
-import net.sourceforge.pmd.annotation.InternalApi;
+package net.sourceforge.pmd.cache.internal;
 
 /**
  * Interface defining an object that has a checksum The checksum is a
@@ -12,11 +10,7 @@ import net.sourceforge.pmd.annotation.InternalApi;
  * changed on the object. It differs from a {@code hashCode()} in that a
  * {@code hashCode()} may not take all fields into account, but a checksum must
  * do so.
- *
- * @deprecated This is internal API, will be hidden with 7.0.0
  */
-@Deprecated
-@InternalApi
 public interface ChecksumAware {
     /**
      * Retrieves the current instance checksum