+++ pmd-core/src/main/java/net/sourceforge/pmd/lang/LanguageProcessor.java
@@ -11,7 +11,7 @@ import org.checkerframework.checker.nullness.qual.NonNull;
 
 import net.sourceforge.pmd.RuleSets;
 import net.sourceforge.pmd.annotation.InternalApi;
-import net.sourceforge.pmd.cache.AnalysisCache;
+import net.sourceforge.pmd.cache.internal.AnalysisCache;
 import net.sourceforge.pmd.lang.document.TextFile;
 import net.sourceforge.pmd.reporting.GlobalAnalysisListener;
 import net.sourceforge.pmd.util.log.MessageReporter;