+++ pmd-core/src/test/java/net/sourceforge/pmd/cache/internal/FileAnalysisCacheTest.java
@@ -2,7 +2,7 @@
  * BSD-style license; for more info see http://pmd.sourceforge.net/license.html
  */
 
-package net.sourceforge.pmd.cache;
+package net.sourceforge.pmd.cache.internal;
 
 import static com.github.stefanbirkner.systemlambda.SystemLambda.restoreSystemProperties;
 import static net.sourceforge.pmd.util.CollectionUtil.listOf;